+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicUserService.java
@@ -3,18 +3,19 @@ package com.sprint.mission.discodeit.service.basic;
 import com.sprint.mission.discodeit.entity.User;
 import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.service.UserService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.UUID;
 
+@Service("basicUserService")
+@RequiredArgsConstructor
 public class BasicUserService implements UserService {
 
     private final UserRepository userRepository;
 
-    public BasicUserService(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
 
     @Override
     public User create(String name, String email, String password) {
@@ -27,7 +28,8 @@ public class BasicUserService implements UserService {
 
     @Override
     public User find(UUID userId) {
-        return userRepository.findById(userId).orElseThrow(() -> new NoSuchElementException("User not found"));
+        return userRepository.findById(userId)
+                .orElseThrow(() -> new NoSuchElementException("User with id " + userId + " not found"));
     }
 
     @Override
@@ -37,13 +39,17 @@ public class BasicUserService implements UserService {
 
     @Override
     public User update(UUID userId, String name, String email, String password) {
-        User user = userRepository.findById(userId).orElseThrow(() -> new NoSuchElementException("User not found"));
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new NoSuchElementException("User with id " + userId + " not found"));
         user.update(name, email, password);
         return userRepository.save(user);
     }
 
     @Override
     public boolean delete(UUID userId) {
+        if (!userRepository.existsById(userId)) {
+            throw new NoSuchElementException("User with id " + userId + " not found");
+        }
         return userRepository.delete(userId);
     }
 }
